Refactor: 변경 - DTOs 변경 및 RequestBody 오류 관리 [#66]
