+++ back/wordseed/src/main/java/com/spring/wordseed/dto/in/CreateCommentInDTO.java
@@ -2,11 +2,12 @@ package com.spring.wordseed.dto.in;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
-@Builder
+@NoArgsConstructor
 public class CreateCommentInDTO {
     private Long postId;
     private String content;
