+++ back/wordseed/src/main/java/com/spring/wordseed/dto/in/CreatePostInDTO.java
@@ -5,11 +5,12 @@ import com.spring.wordseed.enu.PostType;
 import com.spring.wordseed.enu.PostVisibility;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Builder
 @Getter
 @Setter
+@NoArgsConstructor
 public class CreatePostInDTO {
     private String content;
     private String url;
