+++ back/wordseed/src/main/java/com/spring/wordseed/dto/in/DeleteLikeInDTO.java
@@ -1,10 +1,12 @@
-package com.spring.wordseed.controller;
+package com.spring.wordseed.dto.in;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
+@NoArgsConstructor
 public class DeleteLikeInDTO {
     private Long postId;
 }
